src/db_scripts/output_sdk/cisco_nexus_hyperfabric_rest_api_1_0_6_rev_2025_03_11_13103/pyproject.toml
[tool.poetry]
name = "cisco-nexus-hyperfabric-rest-api-client"
version = "1.0.6.post2025.03.11.13103"
description = "A client library for accessing Cisco Nexus Hyperfabric REST API"
authors = []
readme = "README.md"
packages = [
    {include = "cisco_nexus_hyperfabric_rest_api_client"},
]
include = ["CHANGELOG.md", "cisco_nexus_hyperfabric_rest_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
